refactor: [3단계] 구조 및 함수명 리팩터링
