+++ src/main/java/roomescape/auth/AdminInterceptor.java
@@ -29,5 +29,4 @@ public class AdminInterceptor implements HandlerInterceptor {
 		}
 		return true;
 	}
-
 }
