+++ src/main/java/roomescape/auth/LoginMember.java
@@ -1,4 +1,4 @@
-package roomescape.member;
+package roomescape.auth;
 
 public record LoginMember(
 	Long id,
