+++ src/main/java/roomescape/reservation/ReservationController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import roomescape.member.LoginMember;
+import roomescape.auth.LoginMember;
 
 @RestController
 public class ReservationController {
